pyproject.toml
[project]
name = "tds-p2-t32025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "pipreqs>=0.5.0",
    "playwright>=1.56.0",
    "uvicorn>=0.38.0",
]
